wrong version

+++ arch/x86/lib/misc.c
@@ -1,6 +1,4 @@
 // SPDX-License-Identifier: GPL-2.0
-#include <asm/misc.h>
-
 /*
  * Count the digits of @val including a possible sign.
  *
@@ -8,7 +6,7 @@
  */
 int num_digits(int val)
 {
-	long long m = 10;
+	int m = 10;
 	int d = 1;
 
 	if (val < 0) {
